+++ R/class_store.R
@@ -267,15 +267,6 @@ store_unserialize_value <- function(store, target) {
 store_unserialize_value.default <- function(store, target) {
 }
 
-store_get_timestamp <- function(store) {
-  UseMethod("store_get_timestamp")
-}
-
-#' @export
-store_get_timestamp.default <- function(store) {
-  file_info(store$file$path)$mtime
-}
-
 store_validate <- function(store) {
   tar_assert_correct_fields(store, store_new_default)
   store_validate_packages(store)