+++ R/tar_timestamp.R
@@ -1,29 +1,16 @@
 #' @title Get the timestamp(s) of a target.
 #' @export
 #' @family utilities
-#' @description Get the time that a target's data was last modified.
+#' @description Get the time that a target last ran successfully.
 #'   If there are multiple artifacts, as with file or URL targets,
 #'   then multiple time stamps may be returned.
-#' @details `tar_timestamp()` checks the actual data,
-#'   not the metadata, so the returned time stamps
-#'   are more up-to-date than the ones from [tar_meta()].
+#' @details `tar_timestamp()` checks the metadata in `_targets/meta/meta`,
+#'   not the actual data. Time stamps are recorded only for targets that
+#'   run commands: just non-branching targets and individual dynamic
+#'   branches.
 #' @return If the target is not recorded in the metadata
 #'   or cannot be parsed correctly, then
 #'   `tar_timestamp()` returns a `POSIXct` object at `1970-01-01 UTC`.
-#'   If the target is recorded in the metadata and stored locally
-#'   (i.e. the target is not a URL or AWS-backed format) then `tar_timestamp()`
-#'   returns a vector of `POSIXct` modification times of the data files
-#'   (in the order those files are listed in `tar_meta(target_name, path)`).
-#'   The return values for URL and AWS targets depends on the `parse` argument.
-#'   If `parse` is `FALSE`, then `tar_timestamp()` returns an unparsed
-#'   character vector of the time stamps recorded on the web.
-#'   If `parse` is `TRUE`, then `tar_timestamp()` attempts to convert
-#'   those character time stamps into `POSIXct` objects displayed
-#'   with the time zone of the current system. If the time stamp
-#'   cannot be parsed with the given format, `tar_timestamp()`
-#'   returns a `POSIXct` object at `1970-01-01 UTC`
-#'   (so `parse = FALSE` is helpful for debugging
-#'   the `format` argument).
 #' @inheritParams tar_timestamp_raw
 #' @param name Symbol, name of the target. If `NULL` (default)
 #'   then `tar_timestamp()` will attempt to return the timestamp
@@ -55,9 +42,9 @@
 #' }
 tar_timestamp <- function(
   name = NULL,
-  format = "%a, %d %b %Y %H:%M:%S",
-  tz = "UTC",
-  parse = TRUE,
+  format = NULL,
+  tz = NULL,
+  parse = NULL,
   store = targets::tar_config_get("store")
 ) {
   name <- tar_deparse_language(substitute(name))